feat: multi-VM GPU broker (ADR-0007) + MIT/Infinibay licensing
Two VM desktops share one physical GPU with capacity-aware weighted
fair-share and no per-VM license — the VDI differentiator.

infinigpu-sched (new crate, GPU-agnostic, 12 unit tests):
- fail-closed admission + VRAM commit ledger + concurrent-VM cap + per-VM cap
- per-VM token bucket, refill ∝ effective weight (gpuTimeWeight × priority boost)
- weighted fair-share via throttle back-pressure; anti-starvation floor; watchdog
- RAII VmTicket reap (releases VRAM + slot on stop or panic-unwind)

Wired into infinigpu-device: a shared GpuBroker + one SharedGpu Vulkan context
(serialized by the broker run-lock — cooperative, never MPS). InfinigpuBackend
admits at GLOBAL_CTRL enable, routes every render through the scheduler, reaps on
reset. serve_with_broker() serves many VMs off one broker.

Proven on the real A5000 (infinigpu-broker-demo, no QEMU): admission denies over
VRAM ledger / per-VM cap / concurrency cap; a weight-3 designer gets 1.95× the
GPU-time of a weight-1×1.5-boost office desktop (the 2.0× effective-weight target).

Adversarially verified (verify-scheduler workflow: 10 raised, 3 confirmed + fixed):
- a render panic could poison the shared run-lock and brick the whole fleet — now
  contained via catch_unwind + poison-tolerant locks; its guest-reachable trigger
  (unvalidated width×height → u32 overflow) closed by geometry validation + u64
  math. Regression-tested.
- vruntime now uses effective weight (was raw); GPU-context open is pre-warmed off
  the tenant's timed render.

Licensing (per owner): host Rust crates → MIT, Infinibay LLC <andres@infinibay.net>
(+ top-level LICENSE). Guest kernel driver → dual MIT/GPL: the DRM/KMS stack is
EXPORT_SYMBOL_GPL, so a pure-MIT module would be denied those symbols and fail to
load; "Dual MIT/GPL" is the kernel's GPL-compatible identifier that keeps it MIT.

+++ crates/infinigpu-replay/src/lib.rs
@@ -254,7 +254,9 @@ impl HostGpu {
         };
 
         // ---- host-visible readback buffer ----
-        let size = (width * height * 4) as u64;
+        // u64 arithmetic: width*height*4 overflows u32 for large geometries (a debug
+        // build would panic; callers also bound the geometry — verify-scheduler #1).
+        let size = width as u64 * height as u64 * 4;
         let buffer = unsafe {
             dev.create_buffer(
                 &vk::BufferCreateInfo::default()
